server/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
sqlalchemy
sqlmodel
boto3
openai
python-multipart
python-dotenv
httpx
jinja2
beautifulsoup4
html2text
readability-lxml
lxml
sentence-transformers
faiss-cpu
numpy
